move role-functionality to roles.py
